3 files modified, 3 added - react build
