+++ index.html
@@ -1,12 +1,11 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Parcel Sandbox</title>
+    <title>Flow Parcel Sandbox</title>
     <meta charset="UTF-8" />
   </head>
 
   <body>
     <div id="container" style="width:100%;height:100%"></div>
-    <script src="index.js"></script>
   </body>
 </html>
