+++ package.json
@@ -2,18 +2,39 @@
   "name": "flow-diagram-kyrcoh",
   "version": "1.0.0",
   "description": "flow diagram project",
-  "main": "index.html",
-  "scripts": {
-    "start": "parcel index.html --open",
-    "build": "parcel build index.html"
-  },
+  "private": true,
+  "main": "build/index.html",
   "dependencies": {
-    "@antv/g6": "^3.3.2"
+    "@antv/g6": "^3.3.5",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "@babel/core": "7.2.0",
-    "@babel/plugin-proposal-class-properties": "7.8.3",
-    "parcel-bundler": "^1.6.1"
+    "@types/react": "^16.12.0",
+    "@types/react-dom": "^16.12.0",
+    "react-scripts": "3.4.0",
+    "typescript": "^3.8.2"
   },
-  "keywords": []
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  },
+  "keywords": [
+    "javascript",
+    "starter"
+  ],
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
