+++ package.json.bak
@@ -1,0 +1,19 @@
+{
+  "name": "flow-diagram-kyrcoh",
+  "version": "1.0.0",
+  "description": "flow diagram project",
+  "main": "index.html",
+  "scripts": {
+    "start": "parcel index.html --open",
+    "build": "parcel build index.html"
+  },
+  "dependencies": {
+    "@antv/g6": "^3.3.2"
+  },
+  "devDependencies": {
+    "@babel/core": "7.2.0",
+    "@babel/plugin-proposal-class-properties": "7.8.3",
+    "parcel-bundler": "^1.6.1"
+  },
+  "keywords": []
+}
